auth

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,15 @@ use Illuminate\Http\Request;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -52,6 +61,7 @@ class PostsController extends Controller
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = auth()->user()->id;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post created');
@@ -78,6 +88,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return view('/posts')->with('error', 'Unauthorize page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -113,6 +127,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id) {
+            return view('/posts')->with('error', 'Unauthorize page');
+        }
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post removed');
